api-sync/requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
pyodbc
requests
pydantic
python-dateutil